refactor: renamed `collection.dart` to `data.dart`
Renamed `collection.dart` to `data.dart`.
Updated imports to reflect the name change.
Refactored the `visit` function in `data.dart` to use the key for string generation.
Updated `checkProjectExisting` function in `projects.dart` to use `Directory` instead of `File`.
Updated project tile to use opacity to mark if project existing or not.