feat: added project template locale and improved l10n file sorting
Added the ability to define a template locale in project settings.
Updated `WorkingProject` to use the defined template locale.
Improved the sorting of L10n files, placing the template locale at the top.
Enhanced the generation of project short names.
Added a `fileNameMatcher` to the project.
Added the ability to add template locale when creating a project.
Changed locale sorting order.